Make `condval!` parse arbitrary condition exprs
This new approach accumulates tokens for the expression after `if` until
the it matches the pattern `$then:block else $($else:tt)+` for what
follows the expression.

Although this allows for expressions that aren't allowed in normal `if`,
it's better to parse any condition expressions. Such expressions may
also be allowed by normal `if` in the future.